Tidied up the code base
