+++ FF10/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:FF10"
-             StartupUri="MainWindow.xaml">
+             StartupUri="Views\MainWindow.xaml">
     <Application.Resources>
         <BitmapImage x:Key="Open" UriSource="./resource/Open.png"/>
         <BitmapImage x:Key="Save" UriSource="./resource/Save.png"/>
