+++ FF10/Converters/AbilityConverter.cs
@@ -6,18 +6,24 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Data;
 
-namespace FF10
+namespace FF10.Converters
 {
-	class ItemConverter : IValueConverter
+	class AbilityConverter : IValueConverter
 	{
 		public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
 		{
-			return Info.Instance().Search(Info.Instance().Items, (uint)value)?.Name;
+			var abilities = Info.Instance().Abilities;
+			for (int i = 0; i < abilities.Count; i++)
+			{
+				if (abilities[i].Value == (uint)value) return i;
+			}
+
+			return -1;
 		}
 
 		public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
 		{
-			throw new NotImplementedException();
+			return Info.Instance().Abilities[(int)value].Value;
 		}
 	}
 }
