+++ FF10/Converters/BlitzPlayerConverter.cs
@@ -6,18 +6,23 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Data;
 
-namespace FF10
+namespace FF10.Converters
 {
-	class EquipmentConverter : IValueConverter
+	class BlitzPlayerConverter : IValueConverter
 	{
 		public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
 		{
-			return Info.Instance().Search(Info.Instance().Equipments, (uint)value)?.Name;
+			for (int i = 0; i < Info.Instance().Blitz_Player.Count; i++)
+			{
+				if (Info.Instance().Blitz_Player[i].Value == (uint)value) return i;
+			}
+
+			return -1;
 		}
 
 		public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
 		{
-			throw new NotImplementedException();
+			return Info.Instance().Blitz_Player[(int)value].Value;
 		}
 	}
 }
