+++ FF10/Converters/ItemConverter.cs
@@ -6,24 +6,18 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Data;
 
-namespace FF10
+namespace FF10.Converters
 {
-	class AbilityConverter : IValueConverter
+	class ItemConverter : IValueConverter
 	{
 		public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
 		{
-			var abilities = Info.Instance().Abilities;
-			for (int i = 0; i < abilities.Count; i++)
-			{
-				if (abilities[i].Value == (uint)value) return i;
-			}
-
-			return -1;
+			return Info.Instance().Search(Info.Instance().Items, (uint)value)?.Name;
 		}
 
 		public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
 		{
-			return Info.Instance().Abilities[(int)value].Value;
+			throw new NotImplementedException();
 		}
 	}
 }
