+++ FF10/Models/OverDrive.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace FF10
+namespace FF10.Models
 {
 	class OverDrive
 	{
