+++ FF10/Views/AboutWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="FF10.AboutWindow"
+﻿<Window x:Class="FF10.Views.AboutWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:FF10"
+        xmlns:local="clr-namespace:FF10.Views"
         mc:Ignorable="d"
         WindowStartupLocation="CenterScreen"
         Title="AboutWindow" Height="110" Width="250">
