+++ FF10/Views/ChoiceWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="FF10.ChoiceWindow"
+﻿<Window x:Class="FF10.Views.ChoiceWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:FF10"
+        xmlns:local="clr-namespace:FF10.Views"
         mc:Ignorable="d"
         Title="ChoiceWindow" Height="300" Width="300" ResizeMode="NoResize" WindowStartupLocation="CenterScreen"
         Loaded="Window_Loaded">
